Adding information to README on use of Docker container
Removing mod to shebang

This includes a screenshot showing the use of the Docker container, fixes a bug in the tab completion that required a "-" before autocomplete would work, and also removes the dependency on the containerized Python backend.  Users can use the Docker Compose file provided in the HTTP Observatory repo to run the backend locally, or default to the public instance.